refactor: loop routes meta on index for Cards
- minor tweaks to semantic elements and styles
