+++ libs/playground/src/components/Card.tsx
@@ -1,22 +1,9 @@
-import type {
-	HTMLAttributes,
-	ElementType,
-	ReactNode,
-	ImgHTMLAttributes,
-} from 'react'
+import type { ElementType } from 'react'
 import Image from 'next/image'
 import { merge } from '@trikinco/fullstack-components/utils'
 import type { AsComponent } from '@trikinco/fullstack-components'
+import type { CardProps } from '@/src/types/Card'
 
-export interface CardProps
-	extends Omit<HTMLAttributes<HTMLDivElement>, 'title'> {
-	image?: ImgHTMLAttributes<HTMLImageElement>
-	title?: ReactNode
-	children?: ReactNode
-	header?: ReactNode
-	footer?: ReactNode
-	component?: ElementType
-}
 export const defaultElement = 'div'
 
 export const Card = <C extends ElementType = typeof defaultElement>({
@@ -42,7 +29,7 @@ export const Card = <C extends ElementType = typeof defaultElement>({
 	return (
 		<Component
 			className={merge(
-				'relative w-full flex flex-col grow sm:max-w-sm p-4 bg-white border-2 border-gray-200 rounded-lg shadow sm:p-8 dark:bg-slate-900 dark:border-slate-800 focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 focus:ring-offset-slate-50',
+				'relative w-full flex flex-col grow sm:max-w-lg p-4 bg-white border-2 border-gray-200 rounded-lg shadow sm:p-8 dark:bg-slate-900 dark:border-slate-800 focus:outline-none focus:ring-2 focus:ring-slate-400 focus:ring-offset-2 focus:ring-offset-slate-50',
 				className
 			)}
 			{...rest}
@@ -53,11 +40,11 @@ export const Card = <C extends ElementType = typeof defaultElement>({
 				</div>
 			)}
 
-			<div className="flex grow flex-col relative">
+			<div className="flex flex-wrap gap-3 items-center grow sm:flex-nowrap sm:flex-col relative">
 				{src && (
 					<Image
 						className={merge(
-							'mx-auto invert dark:invert-0 hover:mix-blend-hard-light',
+							'w-28 h-28 sm:mx-auto sm:w-auto sm:h-auto invert dark:invert-0 hover:mix-blend-hard-light',
 							imgClassName
 						)}
 						src={src}
@@ -67,15 +54,19 @@ export const Card = <C extends ElementType = typeof defaultElement>({
 					/>
 				)}
 
-				<TitleComponent className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
-					{title}
-				</TitleComponent>
+				<div className="flex flex-col flex-1">
+					<TitleComponent className="mb-2 text-2xl font-bold tracking-tight text-gray-900 dark:text-white">
+						{title}
+					</TitleComponent>
 
-				<div className="mb-3 font-normal text-gray-700 dark:text-gray-300">
-					{children}
+					<div className="mb-3 font-normal text-gray-700 dark:text-gray-300">
+						{children}
+					</div>
+
+					{footer && (
+						<div className="w-full flex gap-2 mt-auto pt-2">{footer}</div>
+					)}
 				</div>
-
-				{footer && <div className="flex gap-2 mt-auto pt-2">{footer}</div>}
 			</div>
 		</Component>
 	)
